This application helps students to prepair their exams related to the computer science subject
This application is a QUIZZ about computer science questions. It allows
students to get ready to the computer science subject in exams.
It has a large database of questions so that it is difficult to repeat
the same question.
having too many stages, you can seem very challenged and keep learning
by playing